[Refactor] task 로직 수정, login 수정
