+++ frontend/src/view/scrum/Task/create/WorkSpaceTaskCreatePage.vue
@@ -1,86 +1,137 @@
 <script setup>
-import { inject, ref, watch, onMounted } from 'vue'; // toRaw 추가
+import { inject, ref, onMounted } from 'vue';
 import { useTaskStore } from '@/stores/scrum/useTaskStore';
 import Multiselect from 'vue-multiselect';
 import { useFriendsStore } from '@/stores/user/useFriendsStore';
 import { useRoute } from 'vue-router';
 import { useSprintStore } from '@/stores/scrum/useSprintStore';
+import { timeInputUtils } from "@/utils/timeInputUtils";
+import { getLabelColors } from "@/utils/labelUtils";
+import { useField, useForm } from 'vee-validate';
+import { Notyf } from 'notyf';
+import * as yup from 'yup';
+import router from "@/router";
+import { useTaskLabelStore } from "@/stores/scrum/useTaskLabelStore";
 
 const contentsTitle = inject('contentsTitle');
 const contentsDescription = inject('contentsDescription');
-contentsTitle.value = 'Work Space Task 추가하기';
-contentsDescription.value = 'Work Space Task를 추가해보세요!';
+contentsTitle.value = '태스크 추가하기';
+contentsDescription.value = '태스크를 추가해보세요!';
 
 const route = useRoute();
 const workspaceId = route.params.workspaceId;
+const notyf = new Notyf();
 
+const friendStore = useFriendsStore();
+const sprintStore = useSprintStore();
 const taskStore = useTaskStore();
-const taskName = ref('');
-const taskContent = ref('');
-const selectedLevel = ref('');
-const selectedPriority = ref('');
+const taskLabelStore = useTaskLabelStore();
+
+const { handleSubmit, errors } = useForm({
+  validationSchema: yup.object({
+    selectedSprintId: yup.string().required('스프린트를 선택해주세요'),
+    taskName: yup.string().required('Task 제목을 입력해주세요'),
+    taskContent: yup.string().required('Task 내용을 입력해주세요'),
+    participants: yup.array().min(1, '담당자를 선택해주세요'),
+    selectedLevel: yup.string().required('난이도를 선택해주세요'),
+    selectedPriority: yup.string().required('중요도를 선택해주세요'),
+    startTime: yup.string().required('시작 날짜를 선택해주세요'),
+    endTime: yup.string().required('종료 날짜를 선택해주세요')
+  })
+});
+
+const { value: selectedSprintId } = useField('selectedSprintId');
+const { value: taskName } = useField('taskName');
+const { value: taskContent } = useField('taskContent');
+const { value: participants } = useField('participants', { initialValue: [] });
+const { value: selectedLevel } = useField('selectedLevel');
+const { value: selectedPriority } = useField('selectedPriority');
+const { value: startTime } = useField('startTime');
+const { value: endTime } = useField('endTime');
+
+const sprintOptions = ref([]);
 const taskSearch = ref('');
 const selectedTasks = ref([]);
-const startTime = ref('');
-const endTime = ref('');
-const selectedSprintId = ref(null);
-const selectedSprint = ref(null);
-const sprintOptions = ref([]);
+const filteredFriends = ref([]);
+const labels = ref([]);
+const availableLabels = ref([]);
+const selectedLabel = ref([]);
+const labelDetails = ref([]);
 
-const friend = useFriendsStore();
-const sprintStore = useSprintStore();
-
-const availableParticipants = ref([]);
-
-const fetchParticipants = async () => {
-  await friend.getFriendsList(workspaceId);
-  if (friend.friends) {
-    availableParticipants.value = friend.friends.map(
-      (friend) => friend.userName
-    );
-  } else {
-    availableParticipants.value = [];
+const searchFriends = async () => {
+  try {
+    await friendStore.getFriendsList(workspaceId);
+    filteredFriends.value = friendStore.friends || [];
+  } catch (error) {
+    console.error('Error fetching Friends:', error);
+    filteredFriends.value = [];
   }
+};
+
+const deleteParticipant = (searchUserIdx) => {
+  participants.value = participants.value.filter(participant => participant.searchUserIdx !== searchUserIdx);
+};
+
+function deleteLabelByName(labelName) {
+  const index = labels.value.findIndex(label => label.labelName === labelName);
+  if (index !== -1) {
+    taskLabelStore.deleteLabel(index);
+    selectedLabel.value = selectedLabel.value.filter(name => name !== labelName);
+    labelDetails.value = labelDetails.value.filter(label => label.labelName !== labelName);
+  }
+}
+
+const onSubmit = handleSubmit(async (values) => {
+  try {
+    const validatedStartTime = timeInputUtils.validateTime(values.startTime);
+    const validatedEndTime = timeInputUtils.validateTime(values.endTime);
+
+    await taskStore.addTask({
+      title: values.taskName,
+      contents: values.taskContent,
+      participants: values.participants.map(participant => participant.searchUserIdx),
+      level: values.selectedLevel,
+      priority: values.selectedPriority,
+      startTime: validatedStartTime,
+      endTime: validatedEndTime,
+      sprintId: selectedSprintId.value
+    });
+
+    taskName.value = '';
+    taskContent.value = '';
+    participants.value = [];
+    selectedLevel.value = '';
+    selectedPriority.value = '';
+    startTime.value = '';
+    endTime.value = '';
+    selectedSprintId.value = null;
+    selectedTasks.value = [];
+    taskSearch.value = '';
+    notyf.success('Task가 추가되었습니다.');
+    router.push(`/workspace/${workspaceId}/scrum/task/list`);
+  } catch (error) {
+    notyf.error('Task 추가에 실패했습니다.');
+    console.error('Error adding task:', error);
+  }
+});
+
+const fetchSprints = async () => {
+  try {
+    await sprintStore.getSprintList(workspaceId);
+    sprintOptions.value = sprintStore.sprints || [];
+  } catch (error) {
+    console.error('Error fetching sprints:', error);
+  }
+};
+const fetchLabels = async () => {
+  await taskLabelStore.getTaskLabel(workspaceId);
+  availableLabels.value = taskLabelStore.labels;
 };
 
 onMounted(async () => {
-  fetchParticipants();
-  await sprintStore.getSprintList(workspaceId);
-  sprintOptions.value = sprintStore.sprints;
-});
-
-watch(selectedSprint, (newSprint) => {
-  if (newSprint) {
-    selectedSprintId.value = newSprint.sprintId; // 선택된 sprintId를 저장
-  } else {
-    selectedSprintId.value = null; // 선택이 해제되면 null로 초기화
-  }
-});
-
-const addTask = () => {
-  const postTask = {
-    sprintId: selectedSprintId.value,
-    title: taskName.value,
-    contents: taskContent.value,
-    startDate: startTime.value,
-    endDate: endTime.value,
-    difficulty: selectedLevel.value,
-    priority: selectedPriority.value,
-    labels: [], // 추가할 라벨 데이터
-    participants: [], // 참여자 추가 (이 부분은 상황에 맞게 처리 필요)
-  };
-
-  taskStore.addTask(postTask, selectedSprintId.value); // 선택된 스프린트 ID로 전달
-  taskName.value = '';
-  selectedLevel.value = '';
-  selectedPriority.value = '';
-  selectedSprintId.value = null;
-};
-
-watch(taskSearch, (newTask) => {
-  if (newTask && !selectedTasks.value.includes(newTask)) {
-    selectedTasks.value.push(newTask);
-  }
+  await searchFriends();
+  await fetchLabels();
+  await fetchSprints();
 });
 </script>
 
@@ -91,95 +142,111 @@ watch(taskSearch, (newTask) => {
         <div>
           <label>스프린트 선택하기</label>
           <multiselect
-            v-model="selectedSprint"
-            :options="sprintOptions"
-            :searchable="true"
-            :close-on-select="true"
-            :show-labels="false"
-            placeholder="스프린트를 선택해주세요"
-            label="title"
-            track-by="sprintId"
+          v-model="selectedSprintId"
+          :options="sprintOptions"
+          :searchable="true"
+          :close-on-select="true"
+          :show-labels="false"
+          placeholder="스프린트를 선택해주세요"
+          label="name"
+          track-by="id"
           />
+          <p class="error-message" v-if="errors.sprintId">{{ errors.sprintId }}</p>
         </div>
         <div>
           <label for="task-name">Task 제목</label>
-          <input
-            type="text"
-            id="task-name"
-            v-model="taskName"
-            placeholder="Task 제목을 적어주세요"
-            class="input-field"
-          />
+          <input type="text" id="task-name" v-model="taskName" placeholder="Task 제목을 적어주세요" class="input-field"/>
+          <p class="error-message" v-if="errors.taskName">{{ errors.taskName }}</p>
         </div>
         <div>
           <label for="task-content">Task 내용</label>
-          <textarea
-            id="task-content"
-            v-model="taskContent"
-            placeholder="Task 내용을 적어주세요"
-            class="input-field"
-            style="margin: 0"
-          />
+          <textarea id="task-content" v-model="taskContent" placeholder="Task 내용을 적어주세요" class="input-field"
+                    style="margin: 0"/>
+          <p class="error-message" v-if="errors.taskContent">{{ errors.taskContent }}</p>
         </div>
         <div>
-          <label>참여자</label>
-          <multiselect
-            v-model="reviewers"
-            :options="availableParticipants"
-            :searchable="true"
-            :close-on-select="true"
-            :show-labels="false"
-            placeholder="담당자를 선택해주세요"
-            label="name"
-            track-by="id"
+          <label>담당자</label>
+          <multiselect v-if="sprintOptions && sprintOptions.length"
+          v-model="selectedSprintId"
+          :options="sprintOptions"
+          :searchable="true"
+          :close-on-select="true"
+          :show-labels="false"
+          placeholder="스프린트를 선택해주세요"
+          label="name"
+          track-by="id"
           />
+          <p class="error-message" v-if="errors.participants">{{ errors.participants }}</p>
+          <div class="selections participants" v-if="participants && participants.length">
+              <span class="item" v-for="participant in participants" :key="participant.searchUserIdx">
+                {{ participant.userName }}
+                <span @click="deleteParticipant(participant.searchUserIdx)"
+                      style="cursor: pointer; margin: 0 10px; padding: 0">x</span>
+              </span>
+          </div>
         </div>
-        <div class="time-wrap">
-          <label>시작 날짜</label>
-          <input
-            v-model="startTime"
-            type="datetime-local"
-            class="time-editor"
-          />
-          <span>~ 종료 날짜</span>
-          <input v-model="endTime" type="datetime-local" class="time-editor" />
+        <div class="inputs-wrap">
+          <div>
+            <div class="time-info">
+              <label>시작 날짜</label>
+              <span>* 시간 지정은 10분 단위로 저장됩니다.</span>
+            </div>
+            <input type="datetime-local" id="startDate" v-model="startTime" class="input-field"/>
+            <p class="error-message" v-if="errors.startDate">{{ errors.startDate }}</p>
+          </div>
+          <div>
+            <div class="time-info">
+              <label>종료 날짜</label>
+              <span>* 시간 지정은 10분 단위로 저장됩니다.</span>
+            </div>
+            <input type="datetime-local" id="endDate" v-model="endTime" class="input-field"/>
+            <p class="error-message" v-if="errors.endDate">{{ errors.endDate }}</p>
+          </div>
         </div>
         <div>
           <label for="level">난이도</label>
           <multiselect
-            v-model="selectedLevel"
-            :options="['LOW', 'MED', 'HIGH']"
-            placeholder="Level"
+              v-model="selectedLevel"
+              :options="['Easy', 'Medium', 'Hard']"
+              placeholder="Level"
           />
+          <p class="error-message" v-if="errors.selectedLevel">{{ errors.selectedLevel }}</p>
         </div>
         <div>
           <label for="priority">중요도</label>
           <multiselect
-            v-model="selectedPriority"
-            :options="['LOW', 'MED', 'HIGH']"
-            placeholder="Priority"
+              v-model="selectedPriority"
+              :options="['Low', 'Medium', 'High']"
+              placeholder="Priority"
           />
+          <p class="error-message" v-if="errors.selectedPriority">{{ errors.selectedPriority }}</p>
         </div>
-        <div class="issue-section">
-          <span class="column">
-            <i class="label-add column-icon"></i>
-            라벨 추가하기
-          </span>
-          <button class="issue-button" @click="rightSideOn('label')">
-            라벨 추가하기
-          </button>
-          <button class="label-button">Frontend</button>
+        <div>
+          <label>라벨 선택</label>
+          <multiselect v-model="selectedLabels" :options="availableLabels || []"></multiselect>
+          <p class="error-message" v-if="errors.sprintLabels">{{ errors.sprintLabels }}</p>
+        </div>
+        <div v-if="selectedLabel" class="label-details">
+          <div v-for="(label, index) in labelDetails" :key="index" class="label-detail-item">
+                <span :style="getLabelColors(label)">
+                  {{ label.labelName }}
+                  <span @click="deleteLabelByName(label.labelName)" style="cursor: pointer; margin: 0 10px; padding: 0">x</span>
+                </span>
+          </div>
         </div>
       </div>
-      <button @click="addTask" class="add-task-btn">Task 연동하기</button>
+      <button @click="onSubmit" class="add-task-btn">Task 추가하기</button>
     </div>
   </div>
 </template>
 
 <style scoped>
 .form-container {
+  margin: 30px;
   box-sizing: border-box;
-  padding: 30px;
+  padding: 20px 40px;
+  border-radius: 10px;
+  background-color: #f7f8fa;
 }
 
 h2 {
@@ -218,6 +285,8 @@ label {
   border: 1px solid #ccc;
   border-radius: 5px;
   font-size: 1rem;
+  background-color: #fff;
+  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.1);
 }
 
 .add-task-btn {
@@ -246,6 +315,19 @@ label {
   margin-bottom: 0;
 }
 
+.add-workspace-btn:hover {
+  background-color: #93AAFD;
+}
+
+.error-message {
+  background-color: #fff1f1;
+  padding: 0.625rem;
+  border-radius: 5px;
+  color: red;
+  font-size: 0.9em;
+  margin: 0;
+}
+
 .time-wrap {
   display: flex;
   gap: 10px;
@@ -258,25 +340,39 @@ label {
   width: 200px;
 }
 
-.issue-button {
-  background-color: #f8d7da;
-  color: #c82333;
-  border: none;
-  padding: 5px 10px;
-  border-radius: 5px;
-  cursor: pointer;
-  margin-right: 10px;
+.time-info {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+
+  span {
+    font-size: 10px;
+    color: #6b7280;
+  }
 }
 
-.issue-id {
-  color: #28303f;
-  background-color: #f3f6ff;
-  padding: 5px 10px;
-  border-radius: 15px;
+.inputs-wrap {
+  display: flex;
+  gap: 20px;
+
+  div {
+    width: 100%;
+  }
+}
+
+.selections {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+  margin-top: 10px;
+}
+
+.item {
   font-size: 12px;
-}
-
-.issue-add {
-  background-image: url('@/assets/icon/boardIcon/issueAdd.svg');
+  font-weight: 500;
+  color: #606C80;
+  padding: 5px 8px;
+  border-radius: 15px;
+  background-color: #e0f7fa;
 }
 </style>
